+++ packages/core/src/actions/account/wallet/create-sub-account.ts
@@ -1,10 +1,10 @@
 import { ManagerClient } from '../../../codegen/abstract'
 import { WithCosmWasmSignOptions } from '../../../types/parameters'
 import { getManagerClientFromApi } from './get-manager-client-from-api'
-import { BaseWalletParameters } from './types'
+import { BaseAccountWalletParameters } from './types'
 
 export type CreateSubAccountParameters = WithCosmWasmSignOptions<
-  BaseWalletParameters &
+  BaseAccountWalletParameters &
     Omit<
       Parameters<typeof ManagerClient.prototype.createSubAccount>[0],
       'accountId'