+++ packages/core/src/actions/account/wallet/get-proxy-client-from-api.ts
@@ -1,8 +1,8 @@
 import { getProxyClient } from '../../wallet/get-proxy-client'
 import { getAccountBaseAddressesFromApi } from '../public/get-account-base-addresses-from-api'
-import { BaseWalletParameters } from './types'
+import { BaseAccountWalletParameters } from './types'
 
-export type GetProxyClientFromApiParameters = BaseWalletParameters
+export type GetProxyClientFromApiParameters = BaseAccountWalletParameters
 
 export async function getProxyClientFromApi({
   accountId,