+++ packages/core/src/actions/account/wallet/propose-owner.ts
@@ -2,10 +2,10 @@ import { ManagerClient } from '../../../codegen/abstract'
 import { Action } from '../../../codegen/abstract/cosmwasm-codegen/Manager.types'
 import { WithCosmWasmSignOptions } from '../../../types/parameters'
 import { getManagerClientFromApi } from './get-manager-client-from-api'
-import { BaseWalletParameters } from './types'
+import { BaseAccountWalletParameters } from './types'
 
 export type ProposeOwnerParameters = WithCosmWasmSignOptions<
-  BaseWalletParameters &
+  BaseAccountWalletParameters &
     Parameters<typeof ManagerClient.prototype.proposeOwner>[0]
 >