+++ packages/core/src/actions/account/wallet/types.ts
@@ -1,7 +1,7 @@
 import { SigningCosmWasmClient } from '@cosmjs/cosmwasm-stargate'
 import { VersionControlTypes } from '../../../codegen/abstract'
 
-export type BaseWalletParameters = {
+export type BaseAccountWalletParameters = {
   accountId: VersionControlTypes.AccountId
   signingCosmWasmClient: SigningCosmWasmClient
   apiUrl: string