+++ packages/core/src/actions/account/wallet/uninstall-module.ts
@@ -2,10 +2,10 @@ import { SigningCosmWasmClient } from '@cosmjs/cosmwasm-stargate'
 import { ManagerClient, VersionControlTypes } from '../../../codegen/abstract'
 import { WithCosmWasmSignOptions } from '../../../types/parameters'
 import { getManagerClientFromApi } from './get-manager-client-from-api'
-import { BaseWalletParameters } from './types'
+import { BaseAccountWalletParameters } from './types'
 
 export type UninstallModulesParameters = WithCosmWasmSignOptions<
-  BaseWalletParameters &
+  BaseAccountWalletParameters &
     Parameters<typeof ManagerClient.prototype.uninstallModule>[0]
 >