+++ packages/core/src/actions/account/wallet/update-info.ts
@@ -1,10 +1,10 @@
 import { ManagerClient } from '../../../codegen/abstract'
 import { WithCosmWasmSignOptions } from '../../../types/parameters'
 import { getManagerClientFromApi } from './get-manager-client-from-api'
-import { BaseWalletParameters } from './types'
+import { BaseAccountWalletParameters } from './types'
 
 export type UpdateInfoParameters = WithCosmWasmSignOptions<
-  BaseWalletParameters &
+  BaseAccountWalletParameters &
     Parameters<typeof ManagerClient.prototype.updateInfo>[0]
 >