+++ packages/core/src/actions/account/wallet/upgrade-modules.ts
@@ -2,14 +2,18 @@ import { SigningCosmWasmClient } from '@cosmjs/cosmwasm-stargate'
 import { ManagerClient, VersionControlTypes } from '../../../codegen/abstract'
 import { WithCosmWasmSignOptions } from '../../../types/parameters'
 import { getManagerClientFromApi } from './get-manager-client-from-api'
-import { BaseWalletParameters } from './types'
+import { BaseAccountWalletParameters } from './types'
 
 export type UpgradeModulesParameters = WithCosmWasmSignOptions<
-  BaseWalletParameters & Parameters<typeof ManagerClient.prototype.upgrade>[0]
+  BaseAccountWalletParameters &
+    Omit<Parameters<typeof ManagerClient.prototype.upgrade>[0], 'accountId'> & {
+      subAccountId?: number
+    }
 >
 
 export async function upgradeModules({
   accountId,
+  subAccountId,
   signingCosmWasmClient,
   apiUrl,
   sender,
@@ -19,7 +23,7 @@ export async function upgradeModules({
   ...rest
 }: UpgradeModulesParameters) {
   const managerClient = await getManagerClientFromApi({
-    accountId,
+    accountId: subAccountId ? { ...accountId, seq: subAccountId } : accountId,
     signingCosmWasmClient,
     sender,
     apiUrl,