+++ examples/wagemos-graz-nextjs/CHANGELOG.md
@@ -1,5 +1,14 @@
 # wagemos-graz-nextjs
 
+## 0.3.19
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @abstract-money/core@1.3.3
+  - @abstract-money/react@1.3.6
+  - @abstract-money/provider-graz@5.0.6
+
 ## 0.3.18
 
 ### Patch Changes